Created classes for converting fetched JSON to objects
