+++ app/src/main/java/nitmeghalaya/shishir2020/model/FacebookPageFeedPaginationCursor.kt
@@ -1,0 +1,10 @@
+package nitmeghalaya.shishir2020.model
+
+/**
+ * Created by Devansh on 8/3/20
+ */
+
+class FacebookPageFeedPaginationCursor(
+    val before: String = "",
+    val after: String = ""
+)
