+++ app/src/main/java/nitmeghalaya/shishir2020/network/RetrofitService.kt
@@ -1,7 +1,9 @@
 package nitmeghalaya.shishir2020.network
 
+import com.squareup.moshi.Moshi
+import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import retrofit2.Retrofit
-import retrofit2.converter.scalars.ScalarsConverterFactory
+import retrofit2.converter.moshi.MoshiConverterFactory
 
 /**
  * Created by Devansh on 8/3/20
@@ -11,9 +13,13 @@ private const val BASE_URL = "https://graph.facebook.com/v6.0/347129762068773/"
 
 object RetrofitService {
 
+    private val moshi = Moshi.Builder()
+        .add(KotlinJsonAdapterFactory())
+        .build()
+
     private val retrofit = Retrofit.Builder()
         .baseUrl(BASE_URL)
-        .addConverterFactory(ScalarsConverterFactory.create())
+        .addConverterFactory(MoshiConverterFactory.create(moshi))
         .build()
 
     fun <T> createService(serviceClass: Class<T>): T = retrofit.create(serviceClass)
