+++ app/src/main/java/nitmeghalaya/shishir2020/repository/FacebookPageRepository.kt
@@ -3,11 +3,13 @@ package nitmeghalaya.shishir2020.repository
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import nitmeghalaya.shishir2020.model.FacebookPageFeed
+import nitmeghalaya.shishir2020.model.FacebookPageFeedItem
 import nitmeghalaya.shishir2020.network.FacebookPageApiService
 import nitmeghalaya.shishir2020.network.RetrofitService
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
+import timber.log.Timber
 
 /**
  * Created by Devansh on 7/3/20
@@ -19,16 +21,16 @@ class FacebookPageRepository {
         RetrofitService.createService(FacebookPageApiService::class.java)
     }
 
-    fun getPageFeed(accessToken: String): LiveData<String> {
-        val pageFeedStringLiveData = MutableLiveData<String>()
+    fun getPageFeed(accessToken: String): LiveData<FacebookPageFeed> {
+        val pageFeedStringLiveData = MutableLiveData<FacebookPageFeed>()
 
-        facebookPageApiService.getPageFeed(accessToken, FacebookPageFeed.getAllProperties())
-            .enqueue(object : Callback<String> {
-                override fun onFailure(call: Call<String>, t: Throwable) {
-                    pageFeedStringLiveData.value = "Failure: " + t.localizedMessage
+        facebookPageApiService.getPageFeed(accessToken, FacebookPageFeedItem.getAllProperties())
+            .enqueue(object : Callback<FacebookPageFeed> {
+                override fun onFailure(call: Call<FacebookPageFeed>, t: Throwable) {
+                    Timber.e("Failure: " + t.localizedMessage)
                 }
 
-                override fun onResponse(call: Call<String>, response: Response<String>) {
+                override fun onResponse(call: Call<FacebookPageFeed>, response: Response<FacebookPageFeed>) {
                     pageFeedStringLiveData.value = response.body()
                 }
             })
